# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Franchise Development OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /permissions:
    get:
      tags: [Access]
      operationId: getPermissionMatrix
      summary: Get the permission catalog and role mapping (manage_permissions)
      responses:
        "200":
          description: Permission matrix
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PermissionMatrix"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /roles/{role}/permissions:
    put:
      tags: [Access]
      operationId: updateRolePermissions
      summary: Replace the permissions granted to a role (manage_permissions)
      parameters:
        - name: role
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UserRole"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RolePermissionsUpdate"
      responses:
        "200":
          description: Updated role mapping
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RolePermissions"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /audit-log:
    get:
      tags: [Access]
      operationId: listAuditLog
      summary: List audit-log entries with optional filters (view_audit_log)
      parameters:
        - name: action
          in: query
          required: false
          schema:
            type: string
        - name: actorId
          in: query
          required: false
          schema:
            type: string
        - name: targetType
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Audit-log entries (newest first)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLogEntry"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /access-requests:
    get:
      tags: [Access]
      operationId: listAccessRequests
      summary: List access/account requests (manage_access_requests)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/AccessRequestStatus"
      responses:
        "200":
          description: Requests (newest first)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccessRequest"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      tags: [Access]
      operationId: createAccessRequest
      summary: >-
        Submit an account or access request. PUBLIC: people without an account
        may submit account requests (rate-limited). Authenticated members may
        also request elevated access.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccessRequestInput"
      responses:
        "201":
          description: Created request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessRequest"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Access (elevation) requests require an authenticated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "429":
          description: Too many requests from this client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /access-requests/{id}/approve:
    post:
      tags: [Access]
      operationId: approveAccessRequest
      summary: Approve a pending request (manage_access_requests)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessRequest"
        "400":
          description: Invalid state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /access-requests/{id}/deny:
    post:
      tags: [Access]
      operationId: denyAccessRequest
      summary: Deny a pending request (manage_access_requests)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Denied request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessRequest"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    UserRole:
      type: string
      enum: [member, lead, admin, vp, super_admin, observer]
...
    PermissionKey:
      type: string
      enum:
        - impersonate_users
        - manage_users
        - manage_permissions
        - view_audit_log
        - manage_access_requests
        - manage_invitations
        - manage_groups
        - assign_tasks
        - approve_completion
        - edit_initiatives
        - delete_initiatives
        - manage_targets
        - view_reports
        - run_reports
        - manage_metrics
        - manage_integrations
    Permission:
      type: object
      required: [key, label, description, category, sortOrder]
      properties:
        key:
          $ref: "#/components/schemas/PermissionKey"
        label:
          type: string
        description:
          type: string
        category:
          type: string
        sortOrder:
          type: integer
    RolePermissions:
      type: object
      required: [role, permissions]
      properties:
        role:
          $ref: "#/components/schemas/UserRole"
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/PermissionKey"
    PermissionMatrix:
      type: object
      required: [catalog, roles]
      description: Permission catalog plus the current role -> permission mapping.
      properties:
        catalog:
          type: array
          items:
            $ref: "#/components/schemas/Permission"
        roles:
          type: array
          items:
            $ref: "#/components/schemas/RolePermissions"
    RolePermissionsUpdate:
      type: object
      required: [permissions]
      properties:
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/PermissionKey"
    AuditLogEntry:
      type: object
      required: [id, action, createdAt]
      properties:
        id:
          type: integer
        actorId:
          type: [string, 'null']
        actorName:
          type: [string, 'null']
        action:
          type: string
        targetType:
          type: [string, 'null']
        targetId:
          type: [string, 'null']
        targetLabel:
          type: [string, 'null']
        before:
          description: Snapshot before the change (shape varies by action).
        after:
          description: Snapshot after the change (shape varies by action).
        createdAt:
          type: string
          format: date-time
    AccessRequestKind:
      type: string
      enum: [account, access]
    AccessRequestStatus:
      type: string
      enum: [pending, approved, denied]
    AccessRequest:
      type: object
      required: [id, kind, requestedRole, status, createdAt]
      properties:
        id:
          type: integer
        kind:
          $ref: "#/components/schemas/AccessRequestKind"
        requesterUserId:
          type: [string, 'null']
        requesterName:
          type: [string, 'null']
        requesterEmail:
          type: [string, 'null']
        requestedRole:
          $ref: "#/components/schemas/UserRole"
        reason:
          type: [string, 'null']
        status:
          $ref: "#/components/schemas/AccessRequestStatus"
        decidedById:
          type: [string, 'null']
        decidedByName:
          type: [string, 'null']
        decidedAt:
          type: [string, 'null']
          format: date-time
        resultInvitationId:
          type: [integer, 'null']
        createdAt:
          type: string
          format: date-time
    AccessRequestInput:
      type: object
      required: [kind]
      description: >-
        Submit an account request (provision a new user) or an access request
        (elevate an existing member). For account requests, name is required.
      properties:
        kind:
          $ref: "#/components/schemas/AccessRequestKind"
        name:
          type: [string, 'null']
        email:
          type: [string, 'null']
          format: email
        requestedRole:
          $ref: "#/components/schemas/UserRole"
        reason:
          type: [string, 'null']
...